# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  date: string;
  createdAt: string;
}

export type CreateTransactionBodyType =
  (typeof CreateTransactionBodyType)[keyof typeof CreateTransactionBodyType];

export const CreateTransactionBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateTransactionBody {
  description: string;
  amount: number;
  type: CreateTransactionBodyType;
  category: string;
  date: string;
}

export interface FinancialSummary {
  totalBalance: number;
  totalIncome: number;
  totalExpenses: number;
  transactionCount: number;
  highestSpendingCategory: string;
  savingsRate: number;
}

export interface CategorySpending {
  category: string;
  amount: number;
  percentage: number;
  count: number;
}

export interface BalanceTrendPoint {
  date: string;
  balance: number;
  income: number;
  expenses: number;
}

export interface MonthlyComparison {
  month: string;
  income: number;
  expenses: number;
  net: number;
}

export type ListTransactionsParams = {
  category?: string;
  type?: ListTransactionsType;
  search?: string;
  sortBy?: ListTransactionsSortBy;
  sortOrder?: ListTransactionsSortOrder;
};

export type ListTransactionsType =
  (typeof ListTransactionsType)[keyof typeof ListTransactionsType];

export const ListTransactionsType = {
  income: "income",
  expense: "expense",
} as const;

export type ListTransactionsSortBy =
  (typeof ListTransactionsSortBy)[keyof typeof ListTransactionsSortBy];

export const ListTransactionsSortBy = {
  date: "date",
  amount: "amount",
  category: "category",
} as const;

export type ListTransactionsSortOrder =
  (typeof ListTransactionsSortOrder)[keyof typeof ListTransactionsSortOrder];

export const ListTransactionsSortOrder = {
  asc: "asc",
  desc: "desc",
} as const;