# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Financial Dashboard API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction operations
  - name: insights
    description: Financial insights
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [date, amount, category]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /insights/summary:
    get:
      operationId: getSummary
      tags: [insights]
      summary: Get financial summary
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"

  /insights/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [insights]
      summary: Get spending breakdown by category
      responses:
        "200":
          description: Category spending breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"

  /insights/balance-trend:
    get:
      operationId: getBalanceTrend
      tags: [insights]
      summary: Get balance trend over time
      responses:
        "200":
          description: Balance trend data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BalanceTrendPoint"

  /insights/monthly-comparison:
    get:
      operationId: getMonthlyComparison
      tags: [insights]
      summary: Get monthly income vs expense comparison
      responses:
        "200":
          description: Monthly comparison data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyComparison"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Transaction:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - description
        - amount
        - type
        - category
        - date
        - createdAt

    CreateTransactionBody:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
          format: date
      required:
        - description
        - amount
        - type
        - category
        - date

    FinancialSummary:
      type: object
      properties:
        totalBalance:
          type: number
        totalIncome:
          type: number
        totalExpenses:
          type: number
        transactionCount:
          type: integer
        highestSpendingCategory:
          type: string
        savingsRate:
          type: number
      required:
        - totalBalance
        - totalIncome
        - totalExpenses
        - transactionCount
        - highestSpendingCategory
        - savingsRate

    CategorySpending:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
        count:
          type: integer
      required:
        - category
        - amount
        - percentage
        - count

    BalanceTrendPoint:
      type: object
      properties:
        date:
          type: string
        balance:
          type: number
        income:
          type: number
        expenses:
          type: number
      required:
        - date
        - balance
        - income
        - expenses

    MonthlyComparison:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        net:
          type: number
      required:
        - month
        - income
        - expenses
        - net