# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all transactions
 */
export const ListTransactionsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  type: zod.enum(["income", "expense"]).optional(),
  search: zod.coerce.string().optional(),
  sortBy: zod.enum(["date", "amount", "category"]).optional(),
  sortOrder: zod.enum(["asc", "desc"]).optional(),
});

export const ListTransactionsResponseItem = zod.object({
  id: zod.number(),
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});
export const ListTransactionsResponse = zod.array(ListTransactionsResponseItem);

/**
 * @summary Create a new transaction
 */
export const CreateTransactionBody = zod.object({
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.coerce.date(),
});

/**
 * @summary Get a transaction by ID
 */
export const GetTransactionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTransactionResponse = zod.object({
  id: zod.number(),
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a transaction
 */
export const UpdateTransactionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTransactionBody = zod.object({
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.coerce.date(),
});

export const UpdateTransactionResponse = zod.object({
  id: zod.number(),
  description: zod.string(),
  amount: zod.number(),
  type: zod.enum(["income", "expense"]),
  category: zod.string(),
  date: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a transaction
 */
export const DeleteTransactionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get financial summary
 */
export const GetSummaryResponse = zod.object({
  totalBalance: zod.number(),
  totalIncome: zod.number(),
  totalExpenses: zod.number(),
  transactionCount: zod.number(),
  highestSpendingCategory: zod.string(),
  savingsRate: zod.number(),
});

/**
 * @summary Get spending breakdown by category
 */
export const GetSpendingByCategoryResponseItem = zod.object({
  category: zod.string(),
  amount: zod.number(),
  percentage: zod.number(),
  count: zod.number(),
});
export const GetSpendingByCategoryResponse = zod.array(
  GetSpendingByCategoryResponseItem,
);

/**
 * @summary Get balance trend over time
 */
export const GetBalanceTrendResponseItem = zod.object({
  date: zod.string(),
  balance: zod.number(),
  income: zod.number(),
  expenses: zod.number(),
});
export const GetBalanceTrendResponse = zod.array(GetBalanceTrendResponseItem);

/**
 * @summary Get monthly income vs expense comparison
 */
export const GetMonthlyComparisonResponseItem = zod.object({
  month: zod.string(),
  income: zod.number(),
  expenses: zod.number(),
  net: zod.number(),
});
export const GetMonthlyComparisonResponse = zod.array(
  GetMonthlyComparisonResponseItem,
);