# lib/api-zod/src/generated/types/balanceTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BalanceTrendPoint {
  date: string;
  balance: number;
  income: number;
  expenses: number;
}