# lib/api-zod/src/generated/types/categorySpending.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySpending {
  category: string;
  amount: number;
  percentage: number;
  count: number;
}