# lib/api-zod/src/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransactionBodyType } from "./createTransactionBodyType";

export interface CreateTransactionBody {
  description: string;
  amount: number;
  type: CreateTransactionBodyType;
  category: string;
  date: Date;
}