# lib/api-zod/src/generated/types/financialSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Financial Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface FinancialSummary {
  totalBalance: number;
  totalIncome: number;
  totalExpenses: number;
  transactionCount: number;
  highestSpendingCategory: string;
  savingsRate: number;
}